producer/requirements.txt
flask==3.0.0
flask-cors==4.0.0
kafka-python==2.0.2